[#19] Refactored server structure.
