+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js",
+    "start": "nodemon server.js",
     "test": "mocha tests/**/*.test.js --exit",
     "test-watch": "nodemon --exec \"npm test\"",
     "lint": "./node_modules/.bin/eslint --ignore-path .gitignore .",
@@ -25,17 +25,13 @@
     "express": "^4.16.4",
     "mongodb": "^3.2.3",
     "mongoose": "^5.5.5",
-    "request": "^2.88.0",
-    "request-promise": "^4.2.4",
     "wikidata-sdk": "^6.4.0"
   },
   "devDependencies": {
     "eslint": "^5.16.0",
     "eslint-config-strongloop": "^2.1.0",
     "expect": "^24.7.1",
-    "jest": "^24.7.1",
     "mocha": "^6.1.4",
-    "nodemon": "^1.18.11",
-    "supertest": "^4.0.2"
+    "nodemon": "^1.18.11"
   }
 }
