+++ routes/routes.js
@@ -1,5 +1,5 @@
 'use strict';
-const controller = require('./controller');
+const controller = require('../handlers/controller');
 /**
  * Exports app routes.
  * This convention enables developers to chain CRUD operations with distinct controllers to
@@ -8,4 +8,5 @@ const controller = require('./controller');
  */
 module.exports = (app) => {
     app.route('/ping').get(controller.ping).post(controller.ping);
+    app.route('/:id').get(controller.getProduct);
 };
